Docker File to create host machine
This file can be used to install an instance of linux (Ubuntu) that can be connected to via RDP. Additionally any changes made to the VM will be saved after shutdown. "start-xrdp.sh" file is installed to the machine to run on start up removing any issues surrounding connecting to the machine after a restart.